remove old TaskRunners
